move json schema route out of runtime group
